lib: Return number of written bytes on flow_write
This is more in line with the write() system call and prepares for
partial writes. Partial writes are disabled by default (and not yet
implemented).

+++ src/tools/ocbr/ocbr_client.c
@@ -135,7 +135,7 @@ int client_main(char * server,
         } else { /* flood */
                 while (!stop) {
                         clock_gettime(CLOCK_REALTIME, &end);
-                        if (flow_write(fd, buf, (size_t) size) < 0) {
+                        if (flow_write(fd, buf, size) < 0) {
                                 stop = true;
                                 continue;
                         }
